% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COMPLEX_WORD_ERROR_PHRASE}
\alias{COMPLEX_WORD_ERROR_PHRASE}
\title{A phrase for detecting complex word errors in soft.check}
\format{
character object
}
\usage{
COMPLEX_WORD_ERROR_PHRASE
}
\description{
A phrase used to detect complex word errors in PNU spell checker.
}
\keyword{datasets}
