% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanspell.R
\docType{class}
\name{hanspell-class}
\alias{hanspell-class}
\alias{hanspell}
\title{hanspell}
\value{

}
\description{
hanspell
}
\section{Slots}{

\describe{
\item{\code{text_corrected}}{corrected text}

\item{\code{text_original}}{original text}

\item{\code{checks}}{a data.table object that provides details on checks performed}
}}

