% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanspell.R
\name{print.hanspell}
\alias{print.hanspell}
\title{Print a hanspell object}
\usage{
\method{print}{hanspell}(object)
}
\arguments{
\item{object}{hanspell object}
}
\value{

}
\description{
Print a hanspell object
}
\examples{
print(spell_check(wrongkortextsample))
}
