% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_check.R
\name{spell_check}
\alias{spell_check}
\title{Spell checker with PNU Korean spell checker (한국어 맞춤법/문법 검사기)}
\usage{
spell_check(text, exceptions = character())
}
\arguments{
\item{text}{Korean text to be checked}

\item{exceptions}{words or phrases that checks contain to be exempted}
}
\value{
a hanspell object that contains
}
\description{
Spell checker with PNU Korean spell checker (한국어 맞춤법/문법 검사기)
}
\examples{
spell_check("왠일이니!")
spell_check("커다란 웃음으로 미안하다 말해야했었지.")
}
