% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_check_daum.R
\name{spell_check_daum}
\alias{spell_check_daum}
\title{Spell checker with Daum Korean spell checker}
\usage{
spell_check_daum(text, exceptions = character())
}
\arguments{
\item{text}{Korean text to be checked}

\item{exceptions}{words or phrases that checks contain to be exempted}
}
\value{
a hanspell object
}
\description{
Spell checker with Daum Korean spell checker
}
\examples{
spell_check_daum(wrongkortextsample)
}
