% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_checker.R
\name{spell_checker}
\alias{spell_checker}
\title{Spell Checker for R}
\usage{
spell_checker(texts)
}
\arguments{
\item{texts}{text for spell check, Korean.}
}
\value{
length 1 character, spell corrected text.
}
\description{
test version
}
\examples{
  spell_checker('留덉땄六愼㸱 寃攼㹣궗')
}
