% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNANALYZABLE_PHRASE_ERROR_PHRASE}
\alias{UNANALYZABLE_PHRASE_ERROR_PHRASE}
\title{A phrase for detecting unanalyzable phrase errors in soft.check}
\format{
character object
}
\usage{
UNANALYZABLE_PHRASE_ERROR_PHRASE
}
\description{
A phrase used to detect unanalyzable phrase errors in PNU spell checker.
}
\keyword{datasets}
