% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wrongkortextsample}
\alias{wrongkortextsample}
\title{A grammatically incorrect Korean sentence.}
\format{
A character with 5 letters.
}
\usage{
wrongkortextsample
}
\description{
A very short grammatically incorrect Korean sentence for tests.
}
\keyword{datasets}
