% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_check.R
\name{spell_check}
\alias{spell_check}
\title{Spell checker with PNU Korean spell checker}
\usage{
spell_check(text, exceptions = character(), soft.check = T)
}
\arguments{
\item{text}{Korean text to be checked}

\item{exceptions}{words or phrases that checks contain to be exempted; default is `character()`}

\item{soft.check}{check if polymorphemic word errors and unanalyzable phrase errors are exempted; default is `TRUE`}
}
\value{
a hanspell object
}
\description{
Spell checker with PNU Korean spell checker
}
\examples{
spell_check(wrongkortextsample)
}
